PR Feedback: Remove uuid from locale test
