+++ processor/src/test/java/org/mapstruct/ap/test/conversion/locale/LocaleConversionTest.java
@@ -3,10 +3,9 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.locale.uuid;
+package org.mapstruct.ap.test.conversion.locale;
 
 import java.util.Locale;
-import java.util.UUID;
 
 import org.mapstruct.ap.testutil.IssueKey;
 import org.mapstruct.ap.testutil.ProcessorTest;
@@ -15,7 +14,7 @@ import org.mapstruct.ap.testutil.WithClasses;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Tests conversions between {@link UUID} and String.
+ * Tests conversions between {@link Locale} and String.
  *
  * @author Jason Bodnar
  */
