+++ processor/src/test/java/org/mapstruct/ap/test/conversion/locale/LocaleMapper.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.locale.uuid;
+package org.mapstruct.ap.test.conversion.locale;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
