+++ processor/src/test/java/org/mapstruct/ap/test/conversion/locale/LocaleSource.java
@@ -3,19 +3,21 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.locale.uuid;
+package org.mapstruct.ap.test.conversion.locale;
+
+import java.util.Locale;
 
 /**
  * @author Jason Bodnar
  */
-public class LocaleTarget {
-    private String localeA;
+public class LocaleSource {
+    private Locale localeA;
 
-    public String getLocaleA() {
+    public Locale getLocaleA() {
         return localeA;
     }
 
-    public void setLocaleA(String localeA) {
+    public void setLocaleA(Locale localeA) {
         this.localeA = localeA;
     }
 }
