+++ processor/src/test/java/org/mapstruct/ap/test/conversion/locale/LocaleTarget.java
@@ -3,21 +3,19 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.locale.uuid;
-
-import java.util.Locale;
+package org.mapstruct.ap.test.conversion.locale;
 
 /**
  * @author Jason Bodnar
  */
-public class LocaleSource {
-    private Locale localeA;
+public class LocaleTarget {
+    private String localeA;
 
-    public Locale getLocaleA() {
+    public String getLocaleA() {
         return localeA;
     }
 
-    public void setLocaleA(Locale localeA) {
+    public void setLocaleA(String localeA) {
         this.localeA = localeA;
     }
 }
